feat: delete article-random
