--- functions/src/interfaces/article-random.ts
@@ -1,6 +1,0 @@
-import { Article } from './article';
-
-export interface ArticleRandom extends Pick<Article, 'articleId' | 'isPublic'> {
-  randomNumber: number;
-  randomCheck: boolean;
-}
