+++ package.json
@@ -4,7 +4,7 @@
   "private": true,
   "license": "MIT",
   "engines": {
-    "node": "12.x"
+    "node": "12"
   },
   "scripts": {
     "start": "ng serve --port 4200 --open",
@@ -14,7 +14,7 @@
     "analyze": "webpack-bundle-analyzer dist/musil/stats.json",
     "test": "ng test",
     "lint": "ng lint",
-    "postinstall": "ngcc"
+    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points"
   },
   "dependencies": {
     "@angular/animations": "~12.2.0",
